modification de l'index.php

+++ blog/index.php
@@ -60,11 +60,19 @@ if (isset($_GET['action'])) {
         $AdminSessionController->logOut();
         
     } else if ($_GET['action'] == 'adminAddPost') {
-        
-        $AdminBackendController->addpost();
+
+        if (!isset($_SESSION["admin_session"])) 
+        {
+            header('Location: index.php?action=logIn');
+            exit;
+        }
+        else
+        {
+            $AdminBackendController->addpost();
+        }
         
     } else if ($_GET['action'] == 'adminEditPost') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["admin_session"])) {
             
             if (isset($_POST['adminEdit'])) {
                 extract($_POST);
@@ -83,48 +91,69 @@ if (isset($_GET['action'])) {
             }
             
         } else {
-            echo 'Erreur';
+             header('Location: index.php?action=logIn');
+             exit;
         }
     } else if ($_GET['action'] == 'adminDeletePost') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["admin_session"])) {
             $AdminBackendController->deletePost($_GET['id']);
         } else {
-            echo 'Erreur';
+            header('Location: index.php?action=logIn');
+            exit;
         }
         
     } else if ($_GET['action'] == 'showComments') {
-        
-        $AdminBackendController->showComments();
-        
+
+        if (!isset($_SESSION["admin_session"])) 
+        {
+            header('Location: index.php?action=logIn');
+            exit;
+        }
+        else
+        {
+            $AdminBackendController->showComments();
+        }
+
     } else if ($_GET['action'] == 'deleteComment') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["admin_session"])) {
             $AdminBackendController->deleteComment($_GET['id']);
         } else {
-            echo 'Erreur';
+             header('Location: index.php?action=logIn');
+             exit;
         }
         
     } else if ($_GET['action'] == 'approveComment') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["admin_session"])) {
             $AdminBackendController->approveComment($_GET['id']);
         } else {
-            echo 'Erreur';
+            header('Location: index.php?action=logIn');
+             exit;
         }
         
     } else if ($_GET['action'] == 'showMembers') {
-        
-        $AdminBackendController->showMembers();
-        
+        if (!isset($_SESSION["admin_session"])) 
+        {
+            header('Location: index.php?action=logIn');
+            exit;
+        }
+        else
+        {
+            $AdminBackendController->showMembers();
+        }
+
     } else if ($_GET['action'] == 'approveMember') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["admin_session"])) {
             $AdminBackendController->approveMember($_GET['id']);
         } else {
-            echo 'Erreur';
+             header('Location: index.php?action=logIn');
+            exit;
         }    
     } else if ($_GET['action'] == 'deleteMember') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["admin_session"])) {
             $AdminBackendController->deleteMember($_GET['id']);
         } else {
-            echo 'Erreur';
+            header('Location: index.php?action=logIn');
+            exit;
         }    
     } else if ($_GET['action'] == 'signIn') {
 
@@ -149,11 +178,18 @@ if (isset($_GET['action'])) {
         $MemberSessionController->signOut();
         
     } else if ($_GET['action'] == 'memberAddPost') {
-        
-        $MemberBackendController->addpost();
-        
+
+         if (!isset($_SESSION["member_session"])) 
+        {
+            header('Location: index.php?action=signIn');
+            exit;
+        }
+        else
+        {
+            $MemberBackendController->addpost();
+        }
     } else if ($_GET['action'] == 'memberEditPost') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["member_session"])) {
             
             if (isset($_POST['memberEdit'])) {
                 extract($_POST);
@@ -172,13 +208,15 @@ if (isset($_GET['action'])) {
             }
             
         } else {
-            echo 'Erreur';
+            header('Location: index.php?action=signIn');
+            exit;
         }
     } else if ($_GET['action'] == 'memberDeletePost') {
-        if (isset($_GET['id']) && $_GET['id'] > 0) {
+        if (isset($_GET['id']) && $_GET['id'] > 0 && isset($_SESSION["member_session"])) {
             $MemberBackendController->deletePost($_GET['id']);
         } else {
-            echo 'Erreur';
+            header('Location: index.php?action=signIn');
+            exit;
         }
     }
 }
